requirements.txt
hidapi>=0.14.0
Pillow>=9.0.0
requests>=2.28.0
